Adding session and upload lengths to request headers of .mparticle requests.

+++ sdk/src/main/java/com/mparticle/ConfigManager.java
@@ -129,7 +129,7 @@ class ConfigManager {
         return localPrefs.mSecret;
     }
 
-    public long getUploadInterval() {
+    public int getUploadInterval() {
         if (localPrefs.debug) {
             return 1000 * localPrefs.debugUploadInterval;
         } else {
